Simplify the API to read package.json files
