--- Nodejs/Product/Npm/IPackageJsonSource.cs
@@ -1,9 +1,0 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-namespace Microsoft.NodejsTools.Npm
-{
-    public interface IPackageJsonSource
-    {
-        dynamic Package { get; }
-    }
-}
